rename catchEvent to catchPromise
