+++ src/events/guildMemberRemove.js
@@ -16,11 +16,11 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 "use strict";
-const catchEvent = require("../utilities/catchEvent.js");
+const catchPromise = require("../utilities/catchPromise.js");
 const client = require("../services/client.js");
 const db = require("../services/database.js");
 
-client.on("guildMemberRemove", catchEvent(async (guild, member) => {
+client.on("guildMemberRemove", catchPromise(async (guild, member) => {
   await db.pool.query(
     "UPDATE users SET in_guild = false WHERE (guild_id, user_id) = ($1, $2)",
     [guild.id, member.id]
